+++ crates/maple_cli/src/stdio_server/session/providers/filer.rs
@@ -91,11 +91,13 @@ impl EventHandler for FilerMessageHandler {
             Event::OnMove(msg) => {
                 #[derive(serde::Deserialize)]
                 struct Params {
-                    curline: String,
+                    // curline: String,
                     cwd: String,
                 }
                 let msg_id = msg.id;
-                let Params { curline, cwd } = msg.deserialize_params_unsafe();
+                // Do not use curline directly.
+                let curline = msg.get_curline(&context.provider_id)?;
+                let Params { cwd } = msg.deserialize_params_unsafe();
                 let path = build_abs_path(&cwd, curline);
                 let on_move_handler = OnMoveHandler {
                     msg_id,
@@ -104,6 +106,7 @@ impl EventHandler for FilerMessageHandler {
                     inner: OnMove::Filer(path.clone()),
                 };
                 if let Err(err) = on_move_handler.handle() {
+                    log::error!("Failed to handle filer OnMove: {:?}", err);
                     let error = json!({"message": err.to_string(), "dir": path});
                     let res = json!({ "id": msg_id, "provider_id": "filer", "error": error });
                     write_response(res);
@@ -145,6 +148,7 @@ pub fn handle_filer_message(msg: Message) {
             json!({ "id": msg.id, "provider_id": "filer", "result": result })
         }
         Err(err) => {
+            log::error!("Failed to read directory entries, cwd: {:?}", cwd);
             let error = json!({"message": err.to_string(), "dir": cwd});
             json!({ "id": msg.id, "provider_id": "filer", "error": error })
         }