+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Make the preview on enter work for `recent_files` provider. #731
 
+## Fixed
+
+- Fix the preview of `filer` provider. #731
+
 ## [0.28] 2021-08-06
 
 ## Improved